move code out of the enormous driver.cpp into more precise locations
