+++ src/AllBuiltNodes.hpp
@@ -4,6 +4,7 @@
 #include "BinaryData.hpp"
 
 struct StatCache;
+struct Driver;
 
 namespace Frozen
 {
@@ -46,3 +47,4 @@ struct AllBuiltNodes
 }
 
 bool OutputFilesMissingFor(const Frozen::BuiltNode* node, StatCache *stat_cache);
+bool SaveAllBuiltNodes(Driver *self);
