+++ src/RemoveStaleOutputs.hpp
@@ -1,0 +1,4 @@
+#pragma once
+struct Driver;
+
+void RemoveStaleOutputs(Driver *self);
